Refacored test
